Change upgrade path to 4.19
