+++ engine/schema/src/main/java/com/cloud/upgrade/dao/Upgrade41800to41810.java
@@ -17,28 +17,17 @@
 package com.cloud.upgrade.dao;
 
 import com.cloud.upgrade.SystemVmTemplateRegistration;
-import com.cloud.utils.crypt.DBEncryptionUtil;
 import com.cloud.utils.exception.CloudRuntimeException;
 import org.apache.log4j.Logger;
-import org.jasypt.exceptions.EncryptionOperationNotPossibleException;
 
 import java.io.InputStream;
 import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 public class Upgrade41800to41810 implements DbUpgrade, DbUpgradeSystemVmTemplate {
     final static Logger LOG = Logger.getLogger(Upgrade41800to41810.class);
     private SystemVmTemplateRegistration systemVmTemplateRegistration;
-
-    static final String ACCOUNT_DETAILS = "account_details";
-
-    static final String DOMAIN_DETAILS = "domain_details";
 
     @Override
     public String[] getUpgradableVersionRange() {
@@ -69,7 +58,6 @@ public class Upgrade41800to41810 implements DbUpgrade, DbUpgradeSystemVmTemplate
     @Override
     public void performDataMigration(Connection conn) {
         fixForeignKeyNames(conn);
-        decryptConfigurationValuesFromAccountAndDomainScopesNotInSecureHiddenCategories(conn);
     }
 
     @Override
@@ -122,57 +110,5 @@ public class Upgrade41800to41810 implements DbUpgrade, DbUpgradeSystemVmTemplate
         DbUpgradeUtils.dropKeysIfExist(conn, "cloud.volumes", keys, true);
         DbUpgradeUtils.dropKeysIfExist(conn, "cloud.volumes", keys, false);
         DbUpgradeUtils.addForeignKey(conn, "volumes", "passphrase_id","passphrase", "id");
-    }
-
-    protected void decryptConfigurationValuesFromAccountAndDomainScopesNotInSecureHiddenCategories(Connection conn) {
-        LOG.info("Decrypting global configuration values from the following tables: account_details and domain_details.");
-
-        Map<Long, String> accountsMap = getConfigsWithScope(conn, ACCOUNT_DETAILS);
-        updateConfigValuesWithScope(conn, accountsMap, ACCOUNT_DETAILS);
-        LOG.info("Successfully decrypted configurations from account_details table.");
-
-        Map<Long, String> domainsMap = getConfigsWithScope(conn, DOMAIN_DETAILS);
-        updateConfigValuesWithScope(conn, domainsMap, DOMAIN_DETAILS);
-        LOG.info("Successfully decrypted configurations from domain_details table.");
-    }
-
-    protected Map<Long, String> getConfigsWithScope(Connection conn, String table) {
-        Map<Long, String> configsToBeUpdated = new HashMap<>();
-        String selectDetails = String.format("SELECT details.id, details.value from cloud.%s details, cloud.configuration c " +
-                "WHERE details.name = c.name AND c.category NOT IN ('Hidden', 'Secure') AND details.value <> \"\" ORDER BY details.id;", table);
-
-        try (PreparedStatement pstmt = conn.prepareStatement(selectDetails)) {
-            try (ResultSet result = pstmt.executeQuery()) {
-                while (result.next()) {
-                    configsToBeUpdated.put(result.getLong("id"), result.getString("value"));
-                }
-            }
-            return configsToBeUpdated;
-        } catch (SQLException e) {
-            String message = String.format("Unable to retrieve data from table [%s] due to [%s].", table, e.getMessage());
-            LOG.error(message, e);
-            throw new CloudRuntimeException(message, e);
-        }
-    }
-
-    protected void updateConfigValuesWithScope(Connection conn, Map<Long, String> configsToBeUpdated, String table) {
-        String updateConfigValues = String.format("UPDATE cloud.%s SET value = ? WHERE id = ?;", table);
-
-        for (Map.Entry<Long, String> config : configsToBeUpdated.entrySet()) {
-            try (PreparedStatement pstmt = conn.prepareStatement(updateConfigValues)) {
-                String decryptedValue = DBEncryptionUtil.decrypt(config.getValue());
-
-                pstmt.setString(1, decryptedValue);
-                pstmt.setLong(2, config.getKey());
-
-                LOG.info(String.format("Updating config with ID [%s] to value [%s].", config.getKey(), decryptedValue));
-                pstmt.executeUpdate();
-            } catch (SQLException | EncryptionOperationNotPossibleException e) {
-                String message = String.format("Unable to update config value with ID [%s] on table [%s] due to [%s]. The config value may already be decrypted.",
-                        config.getKey(), table, e);
-                LOG.error(message);
-                throw new CloudRuntimeException(message, e);
-            }
-        }
     }
 }
